src/main/webapp/app/entities/res-postickets-item/delete/res-postickets-item-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ResPosticketsItemService } from '../service/res-postickets-item.service';

import { ResPosticketsItemDeleteDialogComponent } from './res-postickets-item-delete-dialog.component';

describe('ResPosticketsItem Management Delete Component', () => {
  let comp: ResPosticketsItemDeleteDialogComponent;
  let fixture: ComponentFixture<ResPosticketsItemDeleteDialogComponent>;
  let service: ResPosticketsItemService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [ResPosticketsItemDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(ResPosticketsItemDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ResPosticketsItemDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ResPosticketsItemService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete('ABC');
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith('ABC');
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
